+++ packages/domain-event/src/Doctrine/DomainEventAwareManagerRegistry.php
@@ -17,14 +17,14 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Doctrine\Persistence\ObjectManager;
 use Doctrine\Persistence\ObjectRepository;
-use Rekalogika\DomainEvent\Contracts\DomainEventManagerInterface;
+use Rekalogika\DomainEvent\EventDispatchers;
 use Symfony\Contracts\Service\ResetInterface;
 
 final class DomainEventAwareManagerRegistry extends AbstractManagerRegistryDecorator implements ResetInterface
 {
     public function __construct(
         private ManagerRegistry $wrapped,
-        private DomainEventManagerInterface $domainEventManager,
+        private EventDispatchers $eventDispatchers,
     ) {
         parent::__construct($wrapped);
     }
@@ -51,8 +51,8 @@ final class DomainEventAwareManagerRegistry extends AbstractManagerRegistryDecor
 
         if ($manager instanceof EntityManagerInterface) {
             return new DomainEventAwareEntityManager(
-                $manager,
-                $this->domainEventManager,
+                wrapped: $manager,
+                eventDispatchers: $this->eventDispatchers,
             );
         }
         return $manager;